--- iOS/BanchanCode/BanchanCode/Delegate/CollectionViewConstant.swift
@@ -1,13 +1,0 @@
-//
-//  CollectionViewConstant.swift
-//  BanchanCode
-//
-//  Created by Song on 2021/04/20.
-//
-
-import Foundation
-
-struct CollectionViewConstant {
-    static let numberOfSections: Int = 3
-    static let numberOfItems: Int = 4
-}
