+++ iOS/BanchanCode/BanchanCode/Resource/Assets.xcassets/KeyColors/LaunchBadgeBackgroundColor.colorset/Contents.json
